LayoutTests/platform/win/fast/dom/Element/id-in-deletebutton-expected.txt

Test document.getElementById("WebKit-Editing-Delete-Button")

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL deleteButton should be non-null. Was null
PASS successfullyParsed is true

TEST COMPLETE
1
2
